third_party/blink/web_tests/platform/generic/fast/forms/text/text-window-lost-focus-change-event-expected.txt

Test when window out of focus, should dispatch both blur and change event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS blurEventCounter is 0
PASS changeEventCounter is 0
PASS blurEventCounter is 1
PASS changeEventCounter is 1
PASS successfullyParsed is true

TEST COMPLETE
